Bug #1111884 asoc:sprd:Fix compiling errors
[root cause  ] Build error when open audio config for 9863_32
[changes     ] open Audio config and fix the build error
[side effects] no
[self test   ]:playback
[download normally]:ok
[power on/off normally]:ok
[do common repository/branch inspection]:ok
[is there dependence]:no
[confirm dependent commit]:no
[board]:9863_go_32b
[test case]:playback
[reviewers   ] peter.huang
[change_type ] bringup --> sprd_chip_support
[tag_product ] specific

Function do_div compares the first parameter type: typeof((n)) * to the type:uint64_t *
It caused to the error:comparison of distinct pointer types lacks a cast
So change the type of first parameter

+++ sound/soc/sprd/codec/sprd/sc2721/sprd-headset-sc2721.c
@@ -2663,8 +2663,8 @@ static int sprd_headset_probe(struct platform_device *pdev)
 
 static int headset_adc_get_ideal(u32 adc_mic, u32 coefficient, bool big_scale)
 {
-	int64_t numerator = 0;
-	int64_t denominator = 0;
+	u64 numerator = 0;
+	u64 denominator = 0;
 	u32 adc_ideal = 0;
 	u32 a, b, e1, e2;
 	int64_t exp1, exp2, exp3, exp4;
@@ -2713,7 +2713,7 @@ static int headset_adc_get_ideal(u32 adc_mic, u32 coefficient, bool big_scale)
 			denominator, numerator);
 
 	divisor = (u32)(denominator);
-	dividend = (u64)(numerator);
+	dividend = numerator;
 	pr_info("divisor=%u, dividend=%llu\n", divisor, dividend);
 
 	do_div(dividend, divisor);
